nit: extension method can be moved to perf project
